refactor(runner): colocate the `data:` URI codec in `data-uri.ts`
The matched set was split across files -- `createDataCellURI()` amid the
general link machinery in `link-utils.ts`, `getJSONFromDataURI()` and
`decodeDataURIPayloadText()` amid the entity-URI helpers in
`uri-utils.ts` -- with doc comments doing the work of holding the pair
together. The new `data-uri.ts` contains exactly the codec: the minting
half and the two reading entry points, nothing else. Colocation makes
the pairing structural rather than documentary; the cross-file "matched
set" warnings reduce to a module doc.

Code is moved verbatim (docs updated for the new geography); all
importers rewired. `createDataCellURI` tests move out of
`link-utils.test.ts` into a new `data-uri.test.ts` mirroring the module.
