+++ packages/runner/test/cell-immutable.bench.ts
@@ -1,6 +1,6 @@
 import { Identity } from "@commonfabric/identity";
 import { StorageManager } from "@commonfabric/runner/storage/cache.deno";
-import { createDataCellURI } from "../src/link-utils.ts";
+import { createDataCellURI } from "../src/data-uri.ts";
 import { Runtime } from "../src/runtime.ts";
 import type { IExtendedStorageTransaction } from "../src/storage/interface.ts";
 import type { JSONSchema } from "../src/builder/types.ts";
