+++ packages/runner/test/data-uri-inlining.test.ts
@@ -5,10 +5,8 @@ import { Identity } from "@commonfabric/identity";
 import { StorageManager } from "@commonfabric/runner/storage/cache.deno";
 import { Runtime } from "../src/runtime.ts";
 import type { IExtendedStorageTransaction } from "../src/storage/interface.ts";
-import {
-  createDataCellURI,
-  findAndInlineDataURILinks,
-} from "../src/link-utils.ts";
+import { createDataCellURI } from "../src/data-uri.ts";
+import { findAndInlineDataURILinks } from "../src/link-utils.ts";
 import { LINK_V1_TAG } from "../src/sigil-types.ts";
 
 const signer = await Identity.fromPassphrase("test operator");
