+++ packages/runner/test/schema-links.test.ts
@@ -11,7 +11,8 @@ import type { FabricValue } from "@commonfabric/data-model/fabric-value";
 import { ID, type JSONSchema } from "../src/builder/types.ts";
 import { diffAndUpdate } from "../src/data-updating.ts";
 import { Runtime } from "../src/runtime.ts";
-import { areLinksSame, createDataCellURI } from "../src/link-utils.ts";
+import { createDataCellURI } from "../src/data-uri.ts";
+import { areLinksSame } from "../src/link-utils.ts";
 import { toCell } from "../src/back-to-cell.ts";
 import type { IExtendedStorageTransaction } from "../src/storage/interface.ts";
 import { CellResult } from "../src/query-result-proxy.ts";
